+++ pybop/costs/base_cost.py
@@ -32,6 +32,9 @@ class BaseCost:
     _de : float
         The gradient of the cost function to use if an error occurs during
         evaluation. Defaults to 1.0.
+    minimising : bool, optional, default=True
+        If False, switches the sign of the cost and gradient to perform maximisation
+        instead of minimisation.
     """
 
     def __init__(self, problem: Optional[BaseProblem] = None):
@@ -43,6 +46,7 @@ class BaseCost:
         self.y = None
         self.dy = None
         self._de = 1.0
+        self.minimising = True
         if isinstance(self.problem, BaseProblem):
             self._target = self.problem.target
             self._parameters.join(self.problem.parameters)
@@ -58,6 +62,7 @@ class BaseCost:
         inputs: Union[Inputs, list],
         calculate_grad: bool = False,
         apply_transform: bool = False,
+        for_optimiser: bool = False,
     ) -> Union[float, tuple[float, np.ndarray]]:
         """
         This method calls the forward model via problem.evaluate(inputs),
@@ -73,6 +78,9 @@ class BaseCost:
             cost is computed.
         apply_transform : bool, optional, default=False
             If True, applies a transformation to the inputs before evaluating the model.
+        for_optimiser : bool, optional, default=False
+            If True, returns the cost value if self.minimising=True and the negative of
+            the cost value if self.minimising=False (i.e. the cost is being maximised).
 
         Returns
         -------
@@ -96,6 +104,15 @@ class BaseCost:
         else:
             model_inputs = inputs
 
+        # Check whether we are maximising or minimising via:
+        # | `minimising` | `self.minimising` | `for_optimiser` |
+        # |--------------|-------------------|-----------------|
+        # | `True`       | `True`            | `True`          |
+        # | `True`       | `True`            | `False`         |
+        # | `False`      | `False`           | `True`          |
+        # | `True`       | `False`           | `False`         |
+        minimising = self.minimising or not for_optimiser
+
         # Validate inputs, update parameters
         model_inputs = self.parameters.verify(model_inputs)
         self.parameters.update(values=list(model_inputs.values()))
@@ -110,10 +127,15 @@ class BaseCost:
                     jac = self.transformation.jacobian(inputs)
                     grad = np.matmul(grad, jac)
 
-                return cost, grad
+                return cost * (1 if minimising else -1), grad * (
+                    1 if minimising else -1
+                )
 
             y = self.problem.evaluate(self.problem.parameters.as_dict())
-        return self.compute(y, dy=dy, calculate_grad=calculate_grad)
+
+        return self.compute(y, dy=dy, calculate_grad=calculate_grad) * (
+            1 if minimising else -1
+        )
 
     def compute(self, y: dict, dy: ndarray, calculate_grad: bool = False):
         """